chore: organize files and remove unnecessary constructor fields
