+++ client/src/app/components/audit-template/audit-template-list/audit-template.component.scss
@@ -1,4 +1,4 @@
-@import "./src/theme.scss";
+@import "../../../../theme";
 
 .person-name {
   grid-area: personname;
