+++ client/src/app/components/audit-template/audit-template-list/delete-template-dialog.html
