+++ client/src/app/components/audit-template/audit-template-view/audit-template-view.component.scss
@@ -1,4 +1,4 @@
-@import "./src/theme.scss";
+@import "../../../../theme";
 
 .mat-form-field {
   width: 100%;
