+++ client/src/app/modules/alta-main/alta-main.module.ts
@@ -21,9 +21,9 @@ import { EmployeeSettingsComponent } from '../../components/employee-settings/em
 import { MatMenuModule } from '@angular/material/menu';
 import { MatDialogModule } from '@angular/material/dialog';
 import { ManageInventoryItemsComponent } from '../../components/manage-inventory-items/manage-inventory-items.component';
-import { CreateAuditTemplateComponent } from '../../components/audit-template/create-audit-template/create-audit-template.component';
-import { EditAuditTemplateComponent } from '../../components/audit-template/edit-audit-template/edit-audit-template.component';
-import { AuditTemplateComponent, DeleteTemplateDialogComponent } from '../../components/audit-template/audit-template.component';
+import { CreateAuditTemplateComponent } from '../../components/audit-template/audit-template-view/create-audit-template/create-audit-template.component';
+import { EditAuditTemplateComponent } from '../../components/audit-template/audit-template-view/edit-audit-template/edit-audit-template.component';
+import { AuditTemplateComponent, DeleteTemplateDialogComponent } from '../../components/audit-template/audit-template-list/audit-template.component';
 import {FlexLayoutModule} from '@angular/flex-layout';
 import {MAT_DATE_LOCALE, MatChipsModule, MatDatepickerModule, MatNativeDateModule} from '@angular/material';
 import {MatExpansionModule} from '@angular/material/expansion';
